requirements.txt
amqp==5.1.1
argh==0.27.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==22.12.0
braintree==4.18.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-appconf==1.0.5
django-autoslug==1.9.8
django-celery-email==3.0.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.5
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.8.1
flake8==6.0.0
flower==1.2.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
watchdog==2.2.1
wcwidth==0.2.6
whitenoise==6.3.0
